Doing frontend JS, Error handlering

+++ JS/deleteApplication.script.js
@@ -1,8 +1,18 @@
 const deleteApplication = async function deleteApplication(id) {
-    const response = await fetch(`http://localhost:1000/api/v1/application/deleteApplication/${id}`,{
+    const comformMessage = confirm("Are You Sure You Want to Delete The Appliction")
+     if(!comformMessage){return}
+
+    try{
+        const response = await fetch(`http://localhost:1000/api/v1/application/deleteApplication/${id}`,{
         method:"DELETE"
-    })
-    const data = await response.json()
-    console.log(data);
-    location.reload()
+        })
+        const data = await response.json()
+        console.log(data);
+        location.reload()
+    }
+
+    catch (error) {
+        console.error("Failed to delete:", error);
+        alert("Something went wrong while deleting.");
+    }
 }
